requirements.txt
Flask==3.1.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
PyJWT[crypto]==2.10.1
requests==2.32.3
python-dotenv==1.0.1
cryptography==44.0.2
